voice-agent/requirements.txt
livekit>=1.0.23
livekit-agents>=1.3.12
livekit-plugins-openai>=1.3.12
livekit-plugins-cartesia>=1.3.12
livekit-plugins-silero>=1.3.12
python-dotenv>=1.0.0
anthropic>=0.77.0
